+++ R/controls.R
@@ -252,7 +252,7 @@ subset_taxa_tax_control <-
 
       if (method %in% c("min", "max", "mean", "cutoff_seq")) {
         physeq_seq <- sort(as.vector(as.vector(physeq@otu_table[i, ])))
-        if (sum(physeq_seq > taxa_distri[i])==0) {
+        if (sum(physeq_seq > taxa_distri[i]) == 0) {
           cutoff_seq[i] <- taxa_distri[i]
         } else {
           cutoff_seq[i] <- min(physeq_seq[physeq_seq > taxa_distri[i]])
@@ -291,7 +291,7 @@ subset_taxa_tax_control <-
         0
     }
 
-    n_samp_discard <- sum(taxa_sums(new_physeq) ==0)
+    n_samp_discard <- sum(taxa_sums(new_physeq) == 0)
 
     new_physeq <-
       prune_taxa(colSums(new_physeq@otu_table) > 0, new_physeq)
@@ -299,7 +299,7 @@ subset_taxa_tax_control <-
     nseq_discard <-
       sum(physeq@otu_table) - sum(new_physeq@otu_table)
 
- message(
+    message(
       paste(
         "The filtering processes discard",
         ntaxa_discard,
@@ -307,10 +307,9 @@ subset_taxa_tax_control <-
         nseq_discard,
         "sequences. Note that for ",
         n_samp_discard,
-        "samples, all taxa were discarded. Please run clean_pq() to remove empty samples."
+        "samples, all taxa were discarded. Please run clean_pq() to remove empty samples.",
         sep = " "
       )
-   
     )
     return(new_physeq)
   }