+++ R/plot_functions.R
@@ -1876,28 +1876,29 @@ ggbetween_pq <-
 #' @param clean_pq (logical): Does the phyloseq
 #'   object is cleaned using the [clean_pq()] function?
 #' @param text_size (Num, default 1) A size factor to expand or minimize
-#'   text size. 
+#'   text size.
 #' @param text_size_info (Num, default 1) A size factor to expand or minimize
-#'   text size for extra informations. 
+#'   text size for extra informations.
 #' @examples
 #'
 #' summary_plot_pq(data_fungi)
 #' summary_plot_pq(data_fungi, add_info = FALSE) + scale_fill_viridis_d()
 #' if (requireNamespace("patchwork")) {
-#'   (summary_plot_pq(data_fungi, text_size = 0.5, text_size_info = 0.6) + 
-#'    summary_plot_pq(data_fungi_mini, text_size = 0.5, text_size_info = 0.6)) /
-#'   (summary_plot_pq(data_fungi_sp_known, text_size = 0.5, text_size_info = 0.6) + 
-#'  summary_plot_pq(subset_taxa(data_fungi_sp_known, Phylum=="Ascomycota"),
-#'    text_size = 0.5, text_size_info = 0.6))
+#'   (summary_plot_pq(data_fungi, text_size = 0.5, text_size_info = 0.6) +
+#'     summary_plot_pq(data_fungi_mini, text_size = 0.5, text_size_info = 0.6)) /
+#'     (summary_plot_pq(data_fungi_sp_known, text_size = 0.5, text_size_info = 0.6) +
+#'       summary_plot_pq(subset_taxa(data_fungi_sp_known, Phylum == "Ascomycota"),
+#'         text_size = 0.5, text_size_info = 0.6
+#'       ))
 #' }
 #' @return A ggplot2 object
 #' @export
 summary_plot_pq <- function(physeq,
                             add_info = TRUE,
                             min_seq_samples = 500,
-                            clean_pq = TRUE, 
+                            clean_pq = TRUE,
                             text_size = 1,
-                            text_size_info=1) {
+                            text_size_info = 1) {
   if (clean_pq) {
     physeq <- clean_pq(physeq)
   }
@@ -1973,7 +1974,7 @@ summary_plot_pq <- function(physeq,
         y = y1 + (y2 - y1) / 1.7,
         label = nb_values
       ),
-      size = 4.5*text_size
+      size = 4.5 * text_size
     ) +
     geom_text(
       data = d,
@@ -1982,14 +1983,14 @@ summary_plot_pq <- function(physeq,
         y = y1 + (y2 - y1) / 5,
         label = names
       ),
-      size = 6*text_size
+      size = 6 * text_size
     ) +
     annotate(
       geom = "text",
       x = 0.65,
       y = 3,
       label = "Taxa",
-      size = 6*text_size,
+      size = 6 * text_size,
       color = "#aa4c26",
       angle = 90,
       fontface = 2
@@ -1999,7 +2000,7 @@ summary_plot_pq <- function(physeq,
       x = 0.85,
       y = 3,
       label = "(OTUs, ASVs, ...)",
-      size = 5*text_size,
+      size = 5 * text_size,
       color = "#aa4c26",
       angle = 90
     ) +
@@ -2008,7 +2009,7 @@ summary_plot_pq <- function(physeq,
       x = 2,
       y = 0.65,
       label = "Samples",
-      size = 6*text_size,
+      size = 6 * text_size,
       fontface = 2,
       color = "khaki4"
     )
@@ -2066,7 +2067,7 @@ summary_plot_pq <- function(physeq,
           y = y1 + (y2 - y1) / 2.1,
           label = nb_values
         ),
-        size = 3.5*text_size_info,
+        size = 3.5 * text_size_info,
         hjust = 0
       )
   }
@@ -4988,10 +4989,11 @@ umap_pq <- function(physeq, pkg = "umap", ...) {
 #' plot_complexity_pq(subset_samples(data_fungi_mini, Height == "High"),
 #'   vline_random_kmer = FALSE
 #' )
-#' plot_complexity_pq(subset_samples(data_fungi_mini, Height == "Low"),
-#'   aggregate = FALSE, kmer_size = 4
-#' )
-#' # plot_complexity_pq(subset_samples(data_fungi, Height == "Low"), kmer_size = 4)
+#' # plot_complexity_pq(subset_samples(data_fungi_mini, Height == "Low"),
+#' #  aggregate = FALSE, kmer_size = 4
+#' # )
+#' # plot_complexity_pq(subset_samples(data_fungi, Height == "Low"),
+#' #  kmer_size = 4)
 #'
 #' @details
 #' This function is mainly a wrapper of the work of others.